+++ src/test/java/com/previred/desafio/test/UFValueRetrieverTest.java
@@ -2,17 +2,15 @@ package com.previred.desafio.test;
 
 import static org.junit.Assert.assertNotNull;
 
-import java.time.LocalDate;
-import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.List;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.previred.desafio.UFValueResult;
-import com.previred.desafio.UFValueRetriever;
 import com.previred.desafio.tres.uf.vo.Uf;
+import com.previred.desafio.uf.UFValueResult;
+import com.previred.desafio.uf.UFValueRetriever;
 import com.previred.desafio.util.DateConverter;
 
 public class UFValueRetrieverTest {
